requirements.txt
click
sk-video
gym==0.13
termcolor
flatten_dict
h5py==3.7.0
numpy==1.22.4
torch
mujoco>=2.3.5
dm-control>=1.0.12
Pillow
